int to long long

+++ UniqueValues/unique_finder.cu
@@ -38,7 +38,7 @@ __global__ void count_occurrences_kernel(int* data, int* histogram, int n, int n
 
     // Each thread processes a chunk of data and updates the histogram directly
     for (int i = 0; i < chunk_size; i++) {
-        int dataIdx = globalId * chunk_size + i;
+        long long dataIdx = (long long) globalId * chunk_size + i;
         if (dataIdx < n) {
             atomicAdd(&histogram[data[dataIdx]], 1);
         }
